Extract hover actions and hover contribution point into separate files (#210219)
* extracting actions into a separate file and contribution into a separate file

* updating the imports of the hover controller

* adding correct contribution point in editor.all.ts file
