+++ src/vs/editor/contrib/hover/browser/hoverController.ts
@@ -4,31 +4,21 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { IKeyboardEvent } from 'vs/base/browser/keyboardEvent';
-import { KeyChord, KeyCode, KeyMod } from 'vs/base/common/keyCodes';
+import { KeyCode } from 'vs/base/common/keyCodes';
 import { Disposable, DisposableStore } from 'vs/base/common/lifecycle';
 import { ICodeEditor, IEditorMouseEvent, IPartialEditorMouseEvent, MouseTargetType } from 'vs/editor/browser/editorBrowser';
-import { EditorAction, EditorContributionInstantiation, registerEditorAction, registerEditorContribution, ServicesAccessor } from 'vs/editor/browser/editorExtensions';
 import { ConfigurationChangedEvent, EditorOption } from 'vs/editor/common/config/editorOptions';
 import { Range } from 'vs/editor/common/core/range';
 import { IEditorContribution, IScrollEvent } from 'vs/editor/common/editorCommon';
-import { EditorContextKeys } from 'vs/editor/common/editorContextKeys';
-import { GotoDefinitionAtPositionEditorContribution } from 'vs/editor/contrib/gotoSymbol/browser/link/goToDefinitionAtPosition';
 import { HoverStartMode, HoverStartSource } from 'vs/editor/contrib/hover/browser/hoverOperation';
 import { ContentHoverWidget, ContentHoverController } from 'vs/editor/contrib/hover/browser/contentHover';
 import { MarginHoverWidget } from 'vs/editor/contrib/hover/browser/marginHover';
-import { AccessibilitySupport } from 'vs/platform/accessibility/common/accessibility';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegistry';
-import { editorHoverBorder } from 'vs/platform/theme/common/colorRegistry';
-import { registerThemingParticipant } from 'vs/platform/theme/common/themeService';
-import { HoverParticipantRegistry, IHoverWidget } from 'vs/editor/contrib/hover/browser/hoverTypes';
-import { MarkdownHoverParticipant } from 'vs/editor/contrib/hover/browser/markdownHoverParticipant';
-import { MarkerHoverParticipant } from 'vs/editor/contrib/hover/browser/markerHoverParticipant';
+import { IHoverWidget } from 'vs/editor/contrib/hover/browser/hoverTypes';
 import { InlineSuggestionHintsContentWidget } from 'vs/editor/contrib/inlineCompletions/browser/inlineCompletionsHintsWidget';
 import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
 import { ResultKind } from 'vs/platform/keybinding/common/keybindingResolver';
 import { RunOnceScheduler } from 'vs/base/common/async';
-import * as nls from 'vs/nls';
 import 'vs/css!./hover';
 
 // sticky hover widget which doesn't disappear on focus out and such
@@ -465,429 +455,3 @@ export class HoverController extends Disposable implements IEditorContribution {
 		this._contentWidget?.dispose();
 	}
 }
-
-enum HoverFocusBehavior {
-	NoAutoFocus = 'noAutoFocus',
-	FocusIfVisible = 'focusIfVisible',
-	AutoFocusImmediately = 'autoFocusImmediately'
-}
-
-class ShowOrFocusHoverAction extends EditorAction {
-
-	constructor() {
-		super({
-			id: 'editor.action.showHover',
-			label: nls.localize({
-				key: 'showOrFocusHover',
-				comment: [
-					'Label for action that will trigger the showing/focusing of a hover in the editor.',
-					'If the hover is not visible, it will show the hover.',
-					'This allows for users to show the hover without using the mouse.'
-				]
-			}, "Show or Focus Hover"),
-			metadata: {
-				description: nls.localize2('showOrFocusHoverDescription', 'Show or focus the editor hover which shows documentation, references, and other content for a symbol at the current cursor position.'),
-				args: [{
-					name: 'args',
-					schema: {
-						type: 'object',
-						properties: {
-							'focus': {
-								description: 'Controls if and when the hover should take focus upon being triggered by this action.',
-								enum: [HoverFocusBehavior.NoAutoFocus, HoverFocusBehavior.FocusIfVisible, HoverFocusBehavior.AutoFocusImmediately],
-								enumDescriptions: [
-									nls.localize('showOrFocusHover.focus.noAutoFocus', 'The hover will not automatically take focus.'),
-									nls.localize('showOrFocusHover.focus.focusIfVisible', 'The hover will take focus only if it is already visible.'),
-									nls.localize('showOrFocusHover.focus.autoFocusImmediately', 'The hover will automatically take focus when it appears.'),
-								],
-								default: HoverFocusBehavior.FocusIfVisible,
-							}
-						},
-					}
-				}]
-			},
-			alias: 'Show or Focus Hover',
-			precondition: undefined,
-			kbOpts: {
-				kbExpr: EditorContextKeys.editorTextFocus,
-				primary: KeyChord(KeyMod.CtrlCmd | KeyCode.KeyK, KeyMod.CtrlCmd | KeyCode.KeyI),
-				weight: KeybindingWeight.EditorContrib
-			}
-		});
-	}
-
-	public run(accessor: ServicesAccessor, editor: ICodeEditor, args: any): void {
-		if (!editor.hasModel()) {
-			return;
-		}
-
-		const controller = HoverController.get(editor);
-		if (!controller) {
-			return;
-		}
-
-		const focusArgument = args?.focus;
-		let focusOption = HoverFocusBehavior.FocusIfVisible;
-		if (Object.values(HoverFocusBehavior).includes(focusArgument)) {
-			focusOption = focusArgument;
-		} else if (typeof focusArgument === 'boolean' && focusArgument) {
-			focusOption = HoverFocusBehavior.AutoFocusImmediately;
-		}
-
-		const showContentHover = (focus: boolean) => {
-			const position = editor.getPosition();
-			const range = new Range(position.lineNumber, position.column, position.lineNumber, position.column);
-			controller.showContentHover(range, HoverStartMode.Immediate, HoverStartSource.Keyboard, focus);
-		};
-
-		const accessibilitySupportEnabled = editor.getOption(EditorOption.accessibilitySupport) === AccessibilitySupport.Enabled;
-
-		if (controller.isHoverVisible) {
-			if (focusOption !== HoverFocusBehavior.NoAutoFocus) {
-				controller.focus();
-			} else {
-				showContentHover(accessibilitySupportEnabled);
-			}
-		} else {
-			showContentHover(accessibilitySupportEnabled || focusOption === HoverFocusBehavior.AutoFocusImmediately);
-		}
-	}
-}
-
-class ShowDefinitionPreviewHoverAction extends EditorAction {
-
-	constructor() {
-		super({
-			id: 'editor.action.showDefinitionPreviewHover',
-			label: nls.localize({
-				key: 'showDefinitionPreviewHover',
-				comment: [
-					'Label for action that will trigger the showing of definition preview hover in the editor.',
-					'This allows for users to show the definition preview hover without using the mouse.'
-				]
-			}, "Show Definition Preview Hover"),
-			alias: 'Show Definition Preview Hover',
-			precondition: undefined,
-			metadata: {
-				description: nls.localize2('showDefinitionPreviewHoverDescription', 'Show the definition preview hover in the editor.'),
-			},
-		});
-	}
-
-	public run(accessor: ServicesAccessor, editor: ICodeEditor): void {
-		const controller = HoverController.get(editor);
-		if (!controller) {
-			return;
-		}
-		const position = editor.getPosition();
-
-		if (!position) {
-			return;
-		}
-
-		const range = new Range(position.lineNumber, position.column, position.lineNumber, position.column);
-		const goto = GotoDefinitionAtPositionEditorContribution.get(editor);
-		if (!goto) {
-			return;
-		}
-
-		const promise = goto.startFindDefinitionFromCursor(position);
-		promise.then(() => {
-			controller.showContentHover(range, HoverStartMode.Immediate, HoverStartSource.Keyboard, true);
-		});
-	}
-}
-
-class ScrollUpHoverAction extends EditorAction {
-
-	constructor() {
-		super({
-			id: 'editor.action.scrollUpHover',
-			label: nls.localize({
-				key: 'scrollUpHover',
-				comment: [
-					'Action that allows to scroll up in the hover widget with the up arrow when the hover widget is focused.'
-				]
-			}, "Scroll Up Hover"),
-			alias: 'Scroll Up Hover',
-			precondition: EditorContextKeys.hoverFocused,
-			kbOpts: {
-				kbExpr: EditorContextKeys.hoverFocused,
-				primary: KeyCode.UpArrow,
-				weight: KeybindingWeight.EditorContrib
-			},
-			metadata: {
-				description: nls.localize2('scrollUpHoverDescription', 'Scroll up the editor hover.')
-			},
-		});
-	}
-
-	public run(accessor: ServicesAccessor, editor: ICodeEditor): void {
-		const controller = HoverController.get(editor);
-		if (!controller) {
-			return;
-		}
-		controller.scrollUp();
-	}
-}
-
-class ScrollDownHoverAction extends EditorAction {
-
-	constructor() {
-		super({
-			id: 'editor.action.scrollDownHover',
-			label: nls.localize({
-				key: 'scrollDownHover',
-				comment: [
-					'Action that allows to scroll down in the hover widget with the up arrow when the hover widget is focused.'
-				]
-			}, "Scroll Down Hover"),
-			alias: 'Scroll Down Hover',
-			precondition: EditorContextKeys.hoverFocused,
-			kbOpts: {
-				kbExpr: EditorContextKeys.hoverFocused,
-				primary: KeyCode.DownArrow,
-				weight: KeybindingWeight.EditorContrib
-			},
-			metadata: {
-				description: nls.localize2('scrollDownHoverDescription', 'Scroll down the editor hover.'),
-			},
-		});
-	}
-
-	public run(accessor: ServicesAccessor, editor: ICodeEditor): void {
-		const controller = HoverController.get(editor);
-		if (!controller) {
-			return;
-		}
-		controller.scrollDown();
-	}
-}
-
-class ScrollLeftHoverAction extends EditorAction {
-
-	constructor() {
-		super({
-			id: 'editor.action.scrollLeftHover',
-			label: nls.localize({
-				key: 'scrollLeftHover',
-				comment: [
-					'Action that allows to scroll left in the hover widget with the left arrow when the hover widget is focused.'
-				]
-			}, "Scroll Left Hover"),
-			alias: 'Scroll Left Hover',
-			precondition: EditorContextKeys.hoverFocused,
-			kbOpts: {
-				kbExpr: EditorContextKeys.hoverFocused,
-				primary: KeyCode.LeftArrow,
-				weight: KeybindingWeight.EditorContrib
-			},
-			metadata: {
-				description: nls.localize2('scrollLeftHoverDescription', 'Scroll left the editor hover.'),
-			},
-		});
-	}
-
-	public run(accessor: ServicesAccessor, editor: ICodeEditor): void {
-		const controller = HoverController.get(editor);
-		if (!controller) {
-			return;
-		}
-		controller.scrollLeft();
-	}
-}
-
-class ScrollRightHoverAction extends EditorAction {
-
-	constructor() {
-		super({
-			id: 'editor.action.scrollRightHover',
-			label: nls.localize({
-				key: 'scrollRightHover',
-				comment: [
-					'Action that allows to scroll right in the hover widget with the right arrow when the hover widget is focused.'
-				]
-			}, "Scroll Right Hover"),
-			alias: 'Scroll Right Hover',
-			precondition: EditorContextKeys.hoverFocused,
-			kbOpts: {
-				kbExpr: EditorContextKeys.hoverFocused,
-				primary: KeyCode.RightArrow,
-				weight: KeybindingWeight.EditorContrib
-			},
-			metadata: {
-				description: nls.localize2('scrollRightHoverDescription', 'Scroll right the editor hover.')
-			},
-		});
-	}
-
-	public run(accessor: ServicesAccessor, editor: ICodeEditor): void {
-		const controller = HoverController.get(editor);
-		if (!controller) {
-			return;
-		}
-		controller.scrollRight();
-	}
-}
-
-class PageUpHoverAction extends EditorAction {
-
-	constructor() {
-		super({
-			id: 'editor.action.pageUpHover',
-			label: nls.localize({
-				key: 'pageUpHover',
-				comment: [
-					'Action that allows to page up in the hover widget with the page up command when the hover widget is focused.'
-				]
-			}, "Page Up Hover"),
-			alias: 'Page Up Hover',
-			precondition: EditorContextKeys.hoverFocused,
-			kbOpts: {
-				kbExpr: EditorContextKeys.hoverFocused,
-				primary: KeyCode.PageUp,
-				secondary: [KeyMod.Alt | KeyCode.UpArrow],
-				weight: KeybindingWeight.EditorContrib
-			},
-			metadata: {
-				description: nls.localize2('pageUpHoverDescription', 'Page up the editor hover.'),
-			},
-		});
-	}
-
-	public run(accessor: ServicesAccessor, editor: ICodeEditor): void {
-		const controller = HoverController.get(editor);
-		if (!controller) {
-			return;
-		}
-		controller.pageUp();
-	}
-}
-
-
-class PageDownHoverAction extends EditorAction {
-
-	constructor() {
-		super({
-			id: 'editor.action.pageDownHover',
-			label: nls.localize({
-				key: 'pageDownHover',
-				comment: [
-					'Action that allows to page down in the hover widget with the page down command when the hover widget is focused.'
-				]
-			}, "Page Down Hover"),
-			alias: 'Page Down Hover',
-			precondition: EditorContextKeys.hoverFocused,
-			kbOpts: {
-				kbExpr: EditorContextKeys.hoverFocused,
-				primary: KeyCode.PageDown,
-				secondary: [KeyMod.Alt | KeyCode.DownArrow],
-				weight: KeybindingWeight.EditorContrib
-			},
-			metadata: {
-				description: nls.localize2('pageDownHoverDescription', 'Page down the editor hover.'),
-			},
-		});
-	}
-
-	public run(accessor: ServicesAccessor, editor: ICodeEditor): void {
-		const controller = HoverController.get(editor);
-		if (!controller) {
-			return;
-		}
-		controller.pageDown();
-	}
-}
-
-class GoToTopHoverAction extends EditorAction {
-
-	constructor() {
-		super({
-			id: 'editor.action.goToTopHover',
-			label: nls.localize({
-				key: 'goToTopHover',
-				comment: [
-					'Action that allows to go to the top of the hover widget with the home command when the hover widget is focused.'
-				]
-			}, "Go To Top Hover"),
-			alias: 'Go To Bottom Hover',
-			precondition: EditorContextKeys.hoverFocused,
-			kbOpts: {
-				kbExpr: EditorContextKeys.hoverFocused,
-				primary: KeyCode.Home,
-				secondary: [KeyMod.CtrlCmd | KeyCode.UpArrow],
-				weight: KeybindingWeight.EditorContrib
-			},
-			metadata: {
-				description: nls.localize2('goToTopHoverDescription', 'Go to the top of the editor hover.'),
-			},
-		});
-	}
-
-	public run(accessor: ServicesAccessor, editor: ICodeEditor): void {
-		const controller = HoverController.get(editor);
-		if (!controller) {
-			return;
-		}
-		controller.goToTop();
-	}
-}
-
-
-class GoToBottomHoverAction extends EditorAction {
-
-	constructor() {
-		super({
-			id: 'editor.action.goToBottomHover',
-			label: nls.localize({
-				key: 'goToBottomHover',
-				comment: [
-					'Action that allows to go to the bottom in the hover widget with the end command when the hover widget is focused.'
-				]
-			}, "Go To Bottom Hover"),
-			alias: 'Go To Bottom Hover',
-			precondition: EditorContextKeys.hoverFocused,
-			kbOpts: {
-				kbExpr: EditorContextKeys.hoverFocused,
-				primary: KeyCode.End,
-				secondary: [KeyMod.CtrlCmd | KeyCode.DownArrow],
-				weight: KeybindingWeight.EditorContrib
-			},
-			metadata: {
-				description: nls.localize2('goToBottomHoverDescription', 'Go to the bottom of the editor hover.')
-			},
-		});
-	}
-
-	public run(accessor: ServicesAccessor, editor: ICodeEditor): void {
-		const controller = HoverController.get(editor);
-		if (!controller) {
-			return;
-		}
-		controller.goToBottom();
-	}
-}
-
-registerEditorContribution(HoverController.ID, HoverController, EditorContributionInstantiation.BeforeFirstInteraction);
-registerEditorAction(ShowOrFocusHoverAction);
-registerEditorAction(ShowDefinitionPreviewHoverAction);
-registerEditorAction(ScrollUpHoverAction);
-registerEditorAction(ScrollDownHoverAction);
-registerEditorAction(ScrollLeftHoverAction);
-registerEditorAction(ScrollRightHoverAction);
-registerEditorAction(PageUpHoverAction);
-registerEditorAction(PageDownHoverAction);
-registerEditorAction(GoToTopHoverAction);
-registerEditorAction(GoToBottomHoverAction);
-HoverParticipantRegistry.register(MarkdownHoverParticipant);
-HoverParticipantRegistry.register(MarkerHoverParticipant);
-
-// theming
-registerThemingParticipant((theme, collector) => {
-	const hoverBorder = theme.getColor(editorHoverBorder);
-	if (hoverBorder) {
-		collector.addRule(`.monaco-editor .monaco-hover .hover-row:not(:first-child):not(:empty) { border-top: 1px solid ${hoverBorder.transparent(0.5)}; }`);
-		collector.addRule(`.monaco-editor .monaco-hover hr { border-top: 1px solid ${hoverBorder.transparent(0.5)}; }`);
-		collector.addRule(`.monaco-editor .monaco-hover hr { border-bottom: 0px solid ${hoverBorder.transparent(0.5)}; }`);
-	}
-});
