+++ src/vs/editor/contrib/stickyScroll/browser/stickyScrollModelProvider.ts
@@ -308,7 +308,6 @@ class StickyModelFromCandidateOutlineProvider extends StickyModelCandidateProvid
 			return res;
 		}
 	}
-
 }
 
 abstract class StickyModelFromCandidateFoldingProvider extends StickyModelCandidateProvider<FoldingRegions | null> {
